added ok working piggy

+++ game/src/update.rs
@@ -1,4 +1,4 @@
-use shared::{Context, glam::{Vec2, IVec2}, Grid, Color, Tile};
+use shared::{Context, glam::{Vec2, IVec2, Vec3}, Grid, Color, Tile};
 use crate::{MyGame, Images, sounds};
 use pathfinding::prelude::*;
 
@@ -16,7 +16,7 @@ struct Visit<'a, T:Default + Clone> {
 
 fn cast_ray_mut<T:Default + Clone, F:FnMut(Visit<T>)->bool>(grid:&mut Grid<T>, ray:Ray, mut f:F) {
     fn get_helper(cell_size:f32, pos:f32, dir:f32) -> (f32, f32, f32, f32) {
-        let tile = (pos / cell_size).floor() + 1.0;
+        let tile = (pos / cell_size).floor();// + 1.0;
         let dtile;
         let dt;
         if dir > 0.0 {
@@ -24,7 +24,8 @@ fn cast_ray_mut<T:Default + Clone, F:FnMut(Visit<T>)->bool>(grid:&mut Grid<T>, r
             dt = ((tile + 1.0) * cell_size - pos) / dir;
         } else {
             dtile = -1.0;
-            dt = ((tile - 1.0) * cell_size - pos) / dir;
+            dt = (tile  * cell_size - pos) / dir;
+            // dt = ((tile + 1.0 ) * cell_size - pos) / dir;
         }
 
         (tile, dtile, dt, dtile * cell_size / dir)
@@ -67,6 +68,49 @@ fn cast_ray_mut<T:Default + Clone, F:FnMut(Visit<T>)->bool>(grid:&mut Grid<T>, r
 
 
 impl MyGame {
+
+    fn raycast(&self, start:Vec3, end:Vec3) -> bool {
+        let max_d = (end - start).length();
+        let mut los_blocked = Grid::new(self.state.tiles.size());
+        for (id, door_sprite) in self.state.sprites.iter() {
+            if let Some(door) = self.state.doors.get(id) {
+                if door.open == false {
+                    if let Some(cell) = los_blocked.get_mut(door_sprite.pos.x as i32, door_sprite.pos.y as i32) {
+                        *cell = true;
+                    }
+                }
+            }
+        }
+        let mut was_blocked = false;
+        let f = |visit:Visit<Tile>| {
+            if let Some(blocked) = los_blocked.get(visit.x as i32, visit.y as i32) {
+                if *blocked {
+                    was_blocked = true;
+                    return true;
+                }
+            }
+
+           
+            if visit.tile.clips {
+                was_blocked = true;
+            }
+
+            let d = max_d - visit.d;
+            if d <= 0.0 {
+                return true;
+            }
+
+            visit.tile.clips
+        };
+
+        cast_ray_mut(&mut self.state.tiles.clone(), Ray {
+            start:start.truncate(),
+            end:end.truncate()
+        }, f);
+
+        was_blocked
+    }
+
     fn raycast_update(&mut self, _ctx: &mut dyn Context) {
         let mut los_blocked = Grid::new(self.state.tiles.size());
         for (id, door_sprite) in self.state.sprites.iter() {
@@ -171,19 +215,34 @@ impl MyGame {
             if let Some(player_sprite) = self.state.world.sprites.get(player_id).clone() {
                 for (key, critter_sprite) in self.state.world.sprites.iter_mut() {
                     if let Some(critter) = self.state.critters.get_mut(key) {
-                        let v = (player_sprite.pos - critter_sprite.pos).normalize_or_zero();
+                        let critter_pos = critter_sprite.pos;
                         critter.dir = Vec2::default();
-                        
-                        let path = self.astar(critter_sprite.pos.truncate().as_ivec2(), player_sprite.pos.truncate().as_ivec2());
-                        if let Some(path) = path {
-                            if let Some(first) = path.get(1) {
-                                let p = first.as_vec2() + Vec2::new(0.5, 0.5);
-                                let v = p - critter_sprite.pos.truncate();
-                                critter.dir = v.normalize_or_zero();
-                            } else {
-                                critter.dir = v.truncate().normalize_or_zero();
+
+                        let player_pos = player_sprite.pos;
+                        let los_blocked = self.raycast(critter_pos, player_pos);
+
+                        if los_blocked == false {
+                            critter.last_known_pos = Some(player_pos);
+                        }
+                        if let Some(last_known_pos) = critter.last_known_pos {
+                            let v = last_known_pos - critter_pos;
+                            if v.length() < 0.5 { 
+                                critter.last_known_pos = None;
+                            }
+
+                            //critter.dir = v.truncate().normalize_or_zero();
+                            let path = self.astar(critter_pos.truncate().as_ivec2(), last_known_pos.truncate().as_ivec2());
+                            if let Some(path) = path {
+                                if let Some(first) = path.get(1) {
+                                    let p = first.as_vec2() + Vec2::new(0.5, 0.5);
+                                    let v = p - critter_sprite.pos.truncate();
+                                    critter.dir = v.normalize_or_zero();
+                                } else {
+                                    critter.dir = v.truncate().normalize_or_zero();
+                                }
                             }
                         }
+
                     }
                 }
             }
